requirements.txt
appscript==1.1.2
asgiref==3.3.1
cycler==0.10.0
Django==3.1.7
djangorestframework==3.12.2
docx2pdf==0.1.7
docxtpl==0.11.3
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
Jinja2==2.11.3
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.2
Pillow==8.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tqdm==4.58.0
whitenoise==5.2.0
xlrd==2.0.1
